アニメーションの実装

+++ Animation.cpp
@@ -2,6 +2,11 @@
 
 #include"FbxModel.h"
 
+Animation::Animation()
+{
+
+}
+
 Animation::Animation(float timeTick, float duration)
 {
 	this->timeTick = timeTick;
@@ -10,7 +15,10 @@ Animation::Animation(float timeTick, float duration)
 
 Animation::~Animation()
 {
-	DeleteAnimTree(this->rootNode);
+	if (this->rootNode != nullptr)
+	{
+		DeleteAnimTree(this->rootNode);
+	}
 }
 
 void Animation::DeleteAnimTree(AnimNode* node)
@@ -26,48 +34,57 @@ void Animation::DeleteAnimTree(AnimNode* node)
 	delete node;
 }
 
-void Animation::setFinalTransform(float animationTime, std::vector<glm::mat4>& boneFinalTransforms, AnimNode* node, glm::mat4 parentMatrix, FbxModel* model)
+void Animation::setFinalTransform(float animationTime, std::array<glm::mat4, 251>& boneFinalTransforms, AnimNode* node, glm::mat4 parentMatrix,FbxModel* model)
 {
-	node->getAnimMatrix(animationTime, parentMatrix);
+	glm::mat4 childMatrix = node->getAnimMatrix(animationTime, parentMatrix);
 
 	if (model->containBone(node->getName()))
 	{
-		boneFinalTransforms[model->getBoneToMap(node->getName())] = parentMatrix;
+		int boneToMap = model->getBoneToMap(node->getName());
+
+		if (boneToMap < 251)
+		{
+			boneFinalTransforms[boneToMap] = inverseGlobalTransform * childMatrix * model->getBoneOffset(boneToMap);
+		}
 	}
 
 	for (uint32_t i = 0; i < node->getChildrenCount(); i++)
 	{
-		setFinalTransform(animationTime,boneFinalTransforms, node->getChildren(i), parentMatrix, model);
+		setFinalTransform(animationTime,boneFinalTransforms, node->getChildren(i), childMatrix,model);
 	}
 }
 
-void Animation::setFinalTransform(float animationTime, std::vector<glm::mat4>& boneFinalTransforms, FbxModel* model)
+void Animation::setFinalTransform(float animationTime, std::array<glm::mat4, 251>& boneFinalTransforms,FbxModel* model)
 {
 	glm::mat4 identity(1.0f);
-	rootNode->getAnimMatrix(animationTime,identity);
+	glm::mat4 rootNodeMatrix = rootNode->getAnimMatrix(animationTime,identity);
 
 	if (model->containBone(rootNode->getName()))
 	{
-		boneFinalTransforms[model->getBoneToMap(rootNode->getName())] = identity;
+		int boneToMap = model->getBoneToMap(rootNode->getName());
+		if (boneToMap < 251)
+		{
+			boneFinalTransforms[boneToMap] = inverseGlobalTransform * rootNodeMatrix * model->getBoneOffset(boneToMap);
+		}
 	}
 
 	for (uint32_t i = 0; i < rootNode->getChildrenCount(); i++)
 	{
-		setFinalTransform(animationTime,boneFinalTransforms, rootNode->getChildren(i), identity, model);
+		setFinalTransform(animationTime,boneFinalTransforms, rootNode->getChildren(i), rootNodeMatrix,model);
 	}
+
 }
 
-/*
-std::vector<glm::mat4>& Animation::getBoneTransform()
+
+
+//ポーズの行列をセットする
+void Pose::setPoseMatrix(std::array<glm::mat4, 251>& matrix)
 {
-	animationTime = fmod(timeSeconds * timeTick,duration);
-
-	timeSeconds += 0.1f;
-
-	if (play)
-	{
-		return this->transforms[animationTime];
-	}
-
+	std::copy(matrix.begin(), matrix.end(), boneMatrix.begin());
 }
-*/
+
+//ポーズの行列を返す
+void Pose::setFinalTransform(std::array<glm::mat4, 251>& boneFinalTransforms)
+{
+	std::copy(boneMatrix.begin(), boneMatrix.end(), boneFinalTransforms.begin());
+}
